+++ packages/eyes-sdk-core/lib/sdk/EyesVisualGrid.js
@@ -277,8 +277,8 @@ class EyesVisualGrid extends EyesCore {
   async abort() {
     this._isOpen = false
     return (this._abortPromise = this._abortCommand().then(results => {
-      const resultJson = results.map(result => (result ? {testResults: result.toJSON()} : result)) // not sure if it can even happen that abortCommand from vgc can return partly null array
-      this._runner._allTestResult.push(...resultJson.filter(result => !!result))
+      const resultJson = results.map(result => (result ? result.toJSON() : result)) // not sure if it can even happen that abortCommand from vgc can return partly null array
+      this._runner._allTestResult.push(...resultJson.filter(result => !!result).map(result => ({testResults: result})))
       return resultJson
     }))
   }