+++ packages/screenshoter/src/scroller.js
@@ -25,12 +25,14 @@ function makeScroller({logger, element, scrollingMode = 'mixed'}) {
     if (scrollingMode === 'scroll') return scrollTo(offset, element)
     if (scrollingMode === 'css') return translateTo(offset, element)
     if (scrollingMode === 'mixed') return shiftTo(offset, element)
+    if (scrollingMode === 'mixed+') return shiftTo(offset, element, {formerlyCssScrollingMode: true})
   }
 
   async function getInnerOffset(element = defaultElement) {
     if (scrollingMode === 'scroll') return getScrollOffset(element)
     if (scrollingMode === 'css') return getTranslateOffset(element)
     if (scrollingMode === 'mixed') return getShiftOffset(element)
+    if (scrollingMode === 'mixed+') return getShiftOffset(element)
   }
 
   async function getContentSize() {
@@ -101,9 +103,10 @@ function makeScroller({logger, element, scrollingMode = 'mixed'}) {
     }
   }
 
-  async function shiftTo(offset, element = defaultElement) {
+  async function shiftTo(offset, element = defaultElement, {formerlyCssScrollingMode} = {}) {
     try {
       const scrollOffset = await element.scrollTo(offset)
+      if (utils.geometry.equals(scrollOffset, offset) && !formerlyCssScrollingMode) return scrollOffset
 
       // there is a "bug" in iOS that will not move a root element if it already scrolled, so it should be translated all the way
       if (element.driver.isIOS && (await element.isRoot())) {