+++ packages/snippets/src/browser/getElementRect.js
@@ -1,8 +1,12 @@
 const findFixedAncestor = require('./getElementFixedAncestor')
 const getElementInnerOffset = require('./getElementInnerOffset')
+const getDocumentScrollingElement = require('./getDocumentScrollingElement')
 
 module.exports = function getElementRect([element, isClient = false] = []) {
-  if (element === document.scrollingElement) {
+  if (
+    element === document.documentElement ||
+    element.tagName.toLowerCase() === getDocumentScrollingElement()
+  ) {
     return {x: 0, y: 0, width: element.clientWidth, height: element.clientHeight}
   }