+++ packages/test-utils/src/wait-for-docker-browsers.js
@@ -9,7 +9,10 @@ async function waitForDockerBrowsers({remoteUrl, retries} = {}) {
     )
   if (retries === 0) throw new Error('browsers docker containers failed to start before running tests')
   try {
-    await fetch(remoteUrl)
+    const result = await fetch(remoteUrl)
+    if (result.status === 200) return
+    await delay(300)
+    return waitForDockerBrowsers({remoteUrl, retries: retries - 1})
   } catch (_ex) {
     await delay(300)
     return waitForDockerBrowsers({remoteUrl, retries: retries - 1})